requirements.txt
scapy
openai
pydantic
python-dotenv
